added some things, some other things are broken right now as well
