+++ actions/Actions.js
@@ -1,0 +1,5 @@
+exports.sendMessage = (message, content)  => {
+    message.channel.send(content);
+}
+
+
