--- actions/SendMessage.js
@@ -1,6 +1,0 @@
-function sendMessage(message, response) {
-    message.channel.send(response);
-
-}
-
-module.sendMessage = sendMessage;
