+++ bot.js
@@ -11,17 +11,18 @@ const key = process.env.DISCORD_TOKEN;
 
 client.keychain = {
 	'token' : key,
+	'prefix': process.env.PREFIX,
 }
-const prefix = process.env.PREFIX;
-const owner = process.env.OWNER;
+client.prefix = process.env.PREFIX;
 
 
 
 client.events = eventLoad;
-for (const event in client.events) {
-	console.log(client.events[event].name);
-}
-console.log(client.events);
+client.commands = commandLoad;
+
+
+
+
 client.login(client.keychain['token']).catch(() => {
 	console.error('Login Failed');
 	process.exit(1);
@@ -33,40 +34,3 @@ for (const event in client.events) {
 }
 
 
-client.commands = commandLoad;
-// console.log(client.commands);
-
-client.on('message', (message) => {
-	if (message.content.startsWith(prefix)) {
-
-		//regex for finding the start of the line
-		let command = />(\w+).?([\w\s@#<:>]*)/.exec(message.content);
-
-		let properties = [];
-
-
-		//check if regex passed
-		if (command) {
-			properties.push(message);
-			properties.push(command[1]);
-			//weird regex, check if there is text
-			(!/./.test(command[2]))
-				//push message contents past command
-				properties.push(command[2]);
-		}
-
-		//find command
-
-		let loadedCommand = client.commands[properties[1]];
-		//if exists, run
-		if (loadedCommand) {
-			//super check master if implemented
-			let action = loadedCommand.action(...properties);
-			//if implemented, send back a message
-			if (action) message.channel.send(action);
-		}
-
-
-	}
-
-})
