+++ events/Master.js
@@ -3,6 +3,7 @@ module.exports = class Master {
     constructor(name, type = undefined) {
         this.name = name;
         this.type = type;
+
     }
 
     eventHandler() {
