+++ events/essential/message.js
@@ -7,15 +7,34 @@ module.exports = class Message extends Master {
         )
     }
 
-    eventHandler(message) {
-        if (message.author.bot) return;
-        console.log('received message');
-    }
-
     eventHandler() {
         return function(message) {
+            const client = message.client;
+
             if (message.author.bot) return;
-            message.channel.send('bot under construction :construction:');
+
+            if (!message.content.startsWith(client.keychain['prefix'])) return;
+
+            let command = />(\w+).?([\w\s@#<:>]*)/.exec(message.content);
+            console.log(command);
+            
+            let properties = [];
+            if (command) {
+                properties.push(message);
+                properties.push(command[1]);
+                if (!/./.test(command[2])) 
+                    properties.push(command[2]);
+            }
+
+            let loadedCommand = client.commands[properties[1]];
+
+            if (loadedCommand) {
+                let action = loadedCommand.action(...properties);
+                if (action) message.channel.send(action);
+            }
+
+
+
         }
     }
